requirements.txt
pypdf
chromadb
google-generativeai
pydantic>=2.6
tiktoken
nltk
python-dotenv
streamlit
